backend/requirements.txt
requests~=2.32.4
Django~=5.2.2
djangorestframework~=3.16.0
django-filter~=25.1
djangorestframework-simplejwt~=5.5.0
pillow~=11.3.0
django-cors-headers~=4.7.0
django-stubs~=5.2.1
python-decouple~=3.8
psycopg2-binary~=2.9.10
pytest~=8.4.1
pytest-django~=4.11.1
weasyprint
django-sendfile2